techpack: display: Don't skip LP1 cmd for AoD with fingerprint_mode on
OPlus added extra code to skip sending the LP1/LP2 command if the
fingerprint sensor is active. They probably have other changes they made
in their fork of Android that requires this, but for us, this causes our
panels to not actually enter LP1 mode at all when Ambient Display is
enabled.

Delete special case handling for this so that we can actually use LP1
mode properly.

Change-Id: Ib4a39f14df28ccbf21124defe6283f4a67837bb5
Signed-off-by: mukesh22584 <mks22584@gmail.com>